feat: enforce zero-tolerance compiler warnings and English locale
Enable comprehensive compiler warning detection with -Werror to prevent
technical debt from deprecated APIs, unchecked operations, and other
code quality issues.

## Compiler Configuration Changes

- Enable all compiler warnings: -Xlint:all,-options
- Fail build on any warning: -Werror
- Enforce English locale: .mvn/jvm.config (consistent output across environments)

## Code Quality Fixes

### Fixed Deprecation Warnings (14 instances)
- Replace deprecated @Schema(required=true) with requiredMode=Schema.RequiredMode.REQUIRED
- Files: BatchGraphsRequest, BatchGraphsResponse, CommitResponse, ProblemDetail

### Added Serialization Safety (19 classes)
- Add serialVersionUID to VcException and all 17 subclasses
- Add serialVersionUID to ReadModelProjector.ProjectionException inner class
- Prevents Java serialization warnings and ensures version compatibility

### Removed Redundant Casts (6 test files)
- Remove unnecessary (short) casts for kafkaProperties.getReplicationFactor()
- Method already returns short, cast is redundant per compiler warning

## Documentation Updates

- Document compiler warning enforcement in CLAUDE.md
- Add English locale configuration to build documentation
- Update quality requirements checklist

## Verification

- All 152 source files compile with zero warnings
- Full build passes: mvn clean install
- All ~911 tests pass
- 68MB JAR built successfully

+++ src/test/java/org/chucc/vcserver/controller/BranchDeletionIT.java
@@ -109,7 +109,7 @@ class BranchDeletionIT {
       NewTopic newTopic = new NewTopic(
           topicName,
           kafkaProperties.getPartitions(),
-          (short) kafkaProperties.getReplicationFactor()
+          kafkaProperties.getReplicationFactor()
       );
 
       adminClient.createTopics(Collections.singletonList(newTopic)).all().get();
