+++ src/test/java/org/chucc/vcserver/controller/DatasetDeletionIT.java
@@ -113,7 +113,7 @@ class DatasetDeletionIT {
       NewTopic newTopic = new NewTopic(
           topicName,
           kafkaProperties.getPartitions(),
-          (short) kafkaProperties.getReplicationFactor()
+          kafkaProperties.getReplicationFactor()
       );
 
       adminClient.createTopics(Collections.singletonList(newTopic)).all().get();
