+++ src/test/java/org/chucc/vcserver/projection/ReadModelProjectorIT.java
@@ -103,7 +103,7 @@ class ReadModelProjectorIT {
       NewTopic newTopic = new NewTopic(
           topicName,
           kafkaProperties.getPartitions(),
-          (short) kafkaProperties.getReplicationFactor()
+          kafkaProperties.getReplicationFactor()
       );
 
       adminClient.createTopics(Collections.singletonList(newTopic)).all().get();
